Return serialized person upon GET request to \user
Modify UserServlet.doGet to return a person object serialized in JSON
when a GET request is made to the \user path. The Person class provided
in the models package the ORM .jar file is used to construct the person
object. The Jackson library is used to convert the java object to the
JSON string.